tsc_sync tests: Avoid build and parameterize
Tested by running:
- timerdevice.clock_drift_with_ntp
- timerdevice.tscsync.change_host_clksource
- timerdevice.tscsync.longtime

+++ qemu/tests/timerdevice_clock_drift_with_ntp.py
@@ -26,9 +26,11 @@ def run(test, params, env):
     :param params: Dictionary with test parameters.
     :param env: Dictionary with the test environment.
     """
+    drift_filename = params.get("drift_filename", "/var/lib/ntp/drift")
+
     def _drift_file_exist():
         try:
-            session.cmd("test -f /var/lib/ntp/drift")
+            session.cmd("test -f %s" % drift_filename)
             return True
         except Exception:
             return False
@@ -46,20 +48,25 @@ def run(test, params, env):
     timeout = int(params.get("login_timeout", 360))
     sess_guest_load = vm.wait_for_login(timeout=timeout)
 
-    error.context("Copy time-warp-test.c to guest", logging.info)
-    src_file_name = os.path.join(data_dir.get_deps_dir(), "time_warp",
-                                 "time-warp-test.c")
-    vm.copy_files_to(src_file_name, "/tmp")
+    time_warp_test_cmd = params.get("time_warp_test_cmd", "/tmp/time-warp-test")
 
-    error.context("Compile the time-warp-test.c", logging.info)
-    cmd = "cd /tmp/;"
-    cmd += " yum install -y popt-devel;"
-    cmd += " rm -f time-warp-test;"
-    cmd += " gcc -Wall -o time-warp-test time-warp-test.c -lrt"
-    sess_guest_load.cmd(cmd)
+    if sess_guest_load.get_command_status("test -x %s" % time_warp_test_cmd):
+        error.context("Copy time-warp-test.c to guest", logging.info)
+        src_file_name = os.path.join(data_dir.get_deps_dir("tsc_sync"),
+                                     "time-warp-test.c")
+        vm.copy_files_to(src_file_name, "/tmp")
+
+        error.context("Compile the time-warp-test.c", logging.info)
+        cmd = "cd /tmp/;"
+        cmd += " yum install -y popt-devel;"
+        cmd += " rm -f time-warp-test;"
+        cmd += " gcc -Wall -o time-warp-test time-warp-test.c -lrt"
+        sess_guest_load.cmd(cmd)
 
     error.context("Stop ntpd and apply load on guest", logging.info)
-    sess_guest_load.cmd("yum install -y ntp; service ntpd stop")
+    default_ntp_stop_cmd = "yum install -y ntp; service ntpd stop; rm -f %s" % drift_filename
+    ntp_stop_cmd = params.get("ntp_stop_cmd", default_ntp_stop_cmd)
+    sess_guest_load.cmd(ntp_stop_cmd)
     load_cmd = "for ((I=0; I<`grep 'processor id' /proc/cpuinfo| wc -l`; I++));"
     load_cmd += " do taskset $(( 1 << $I )) /bin/bash -c 'for ((;;)); do X=1; done &';"
     load_cmd += " done"
@@ -80,12 +87,12 @@ def run(test, params, env):
 
     error.context("Run time-warp-test", logging.info)
     session = vm.wait_for_login(timeout=timeout)
-    cmd = "/tmp/time-warp-test > /dev/null &"
+    cmd = "%s > /dev/null &" % time_warp_test_cmd
     session.sendline(cmd)
 
     error.context("Start ntpd on guest", logging.info)
-    cmd = "service ntpd start; sleep 1; echo"
-    session.cmd(cmd)
+    ntp_start_cmd = params.get("ntp_start_cmd", "service ntpd start; sleep 1; echo")
+    session.cmd(ntp_start_cmd)
 
     error.context("Check if the drift file exists on guest", logging.info)
     test_run_timeout = float(params["test_run_timeout"])
@@ -93,10 +100,12 @@ def run(test, params, env):
         utils_misc.wait_for(_drift_file_exist, test_run_timeout, step=5)
     except aexpect.ShellCmdError, detail:
         raise error.TestError("Failed to wait for the creation of"
-                              " /var/lib/ntp/drift file. Detail: '%s'" % detail)
+                              " %s file. Detail: '%s'" %
+                              (drift_filename, detail))
 
     error.context("Verify the drift file content on guest", logging.info)
-    output = session.cmd("cat /var/lib/ntp/drift")
+    output = session.cmd("cat %s" % drift_filename)
     if int(abs(float(output))) > 20:
         raise error.TestFail("Failed to check the ntp drift."
-                             " Output: '%s'" % output)
+                             " Output from 'cat %s': '%s'" %
+                             (drift_filename, output))
